Rename slr package to simple
