+++ parser/lr/simple/closure_test.go
@@ -1,4 +1,4 @@
-package slr
+package simple
 
 import (
 	"testing"
