+++ parser/lr/simple/item.go
@@ -1,4 +1,4 @@
-package slr
+package simple
 
 import (
 	"bytes"
